+++ samples/WeihanLi.Redis.Samples/Program.cs
@@ -1,5 +1,6 @@
-﻿using System;
-using Microsoft.Extensions.DependencyInjection;
+﻿using Microsoft.Extensions.DependencyInjection;
+using StackExchange.Redis;
+using System;
 using WeihanLi.Common;
 using WeihanLi.Common.Helpers;
 using WeihanLi.Common.Logging.Log4Net;
@@ -10,21 +11,19 @@ namespace WeihanLi.Redis.Samples
     {
         public static void Main(string[] args)
         {
-            LogHelper.AddLogProvider(new Log4NetLogHelperProvider());
+            LogHelper.LogFactory.AddLog4Net();
 
             IServiceCollection services = new ServiceCollection();
             services.AddRedisConfig(options =>
             {
-                options.EnableCompress = true;
                 options.RedisServers = new[]
                 {
-                    //new RedisServerConfiguration("127.0.0.1"),
-                    new RedisServerConfiguration("127.0.0.1", 16379),
+                    new RedisServerConfiguration("127.0.0.1"),
+                    //new RedisServerConfiguration("127.0.0.1", 16379),
                 };
             });
             // custom serializer
-            services.AddSingleton<IDataSerializer, BinaryDataSerializer>();
-            services.AddSingleton<IDataCompressor, NullDataCompressor>();
+            //services.AddSingleton<IDataSerializer, BinaryDataSerializer>();
             DependencyResolver.SetDependencyResolver(services);
 
             //var cacheClient = DependencyResolver.Current.ResolveService<ICacheClient>();
@@ -39,7 +38,38 @@ namespace WeihanLi.Redis.Samples
             //var result = cacheClient.Get<PagedListModel<int>>(customSerializerCacheKey);
             //Console.WriteLine(result.ToJson());
 
-            ConfigurationChangedEventSample.MainTest();
+            var database = DependencyResolver.Current.GetRequiredService<IConnectionMultiplexer>().GetDatabase();
+            var c_name = "test_counter";
+            database.StringSet(c_name, 0, TimeSpan.FromSeconds(10));
+
+            var val = database.StringDecrement(c_name);
+            Console.WriteLine(val);
+            val = database.StringIncrement(c_name);
+            Console.WriteLine(val);
+
+            //try
+            //{
+            //    var cts = new CancellationTokenSource();
+            //    var task = Task.Delay(3000, cts.Token);
+            //    var task2 = Task.Delay(1000);
+
+            //    cts.Cancel(true);
+
+            //    Thread.Sleep(1000);
+
+            //    Console.WriteLine($"task.IsCompleted:{task.IsCompleted}, task.IsCanceled:{task.IsCanceled}");
+            //    Console.WriteLine($"task2.IsCompleted:{task2.IsCompleted}, task2.IsCanceled:{task2.IsCanceled}");
+            //}
+            //catch (TaskCanceledException ex)
+            //{
+            //    Console.WriteLine($"task canceled, ex:{ex}");
+            //}
+            //catch (Exception e)
+            //{
+            //    Console.WriteLine(e);
+            //}
+
+            //ConfigurationChangedEventSample.MainTest();
 
             Console.ReadLine();
         }